Change CrwJsonRpc to accept session and user_id
CrwJsonRpc now extends the JsonRpcServer (and the json rpc 2.0 protocol)
to accept two additional parameters in the request:
 * 'user_id': this should be an integer and it should be the user_id of
 the user that is sending the request.
 * 'session': this should be a string and should be the session key of
 the user that is sending the request.

Both parameters are optional for unauthenticated requests, but are
required for any request that is only allowed for logged in users. For
example, login and create_account don't require these parameters, but
create_team and add_to_team do.

This change centralises the verification of the user, so that less code
needs to be repeated and security bugs (like mistakingly verifying an
user) are less likely. Furthermore, it streamlines the rpc request
workflow for the client, since they don't need to add these things in
the parameters. This can now be handled transparently in an
create_request function on the client.

To extend the protocol in CrwJsonRpc, the rpc_invoke_single command
needed to be overridden. However, this was a private method in
JsonRpcServer, to override it, it has been moved to be an actual method
of JsonRpcServer. This method has then been overridden in CrwJsonRpc.

This commit also rewrites the tests in test_crw_jsonrpc, to work with
the new format. Some tests needed to be changed to send an actual
request and parse a response, since the authentication is now handled by
the JsonRpcServer.rpc_invoke method (or more accurately, by the
overriden rpc_invoke_single method in CrwJsonRpc).