Rename promo site images to lowercase
Because the promo website was created on Microsoft Windows, where
filenames are case insensitive, some images on the site technically had
parts of the filename in uppercase, while being refered to entirely in
lowercase. This caused problems when the site was hosted on GNU/Linux,
which uses a case sensitive file system, causing some of the images to
not be found. This commit renames all the image files which contain
upper case characters to the same name in lower case, fixing this
issue.