Find user_id implicitly when it isn't passed
When an RPC request misses an user_id or has an invalid user_id, the
server now finds the user_id that belongs to the session_key if it is
given and uses that id. This way the client doesn't have to save the
user_id. However, this greatly reduces the security of the sessions :/.